refactor code for getting the number of features
