+++ onmt/inputters/audio_dataset.py
@@ -192,26 +192,3 @@ class AudioDataset(DatasetBase):
                 index += 1
 
                 yield example_dict
-
-    @staticmethod
-    def get_num_features(corpus_file, side):
-        """
-        For audio corpus, source side is in form of audio, thus
-        no feature; while target side is in form of text, thus
-        we can extract its text features.
-
-        Args:
-            corpus_file (str): file path to get the features.
-            side (str): 'src' or 'tgt'.
-
-        Returns:
-            number of features on `side`.
-        """
-        if side == 'src':
-            num_feats = 0
-        else:
-            with codecs.open(corpus_file, "r", "utf-8") as cf:
-                f_line = cf.readline().strip().split()
-                _, _, num_feats = AudioDataset.extract_text_features(f_line)
-
-        return num_feats
